:recycle: refactors test_main into multiple files
